Move things around to make adding new apps easier
