+++ .metadata.sh
@@ -11,7 +11,7 @@ HBREPO="golift/homebrew-mugs"
 MAINT="David Newhall II <david at sleepers dot pro>"
 VENDOR="Go Lift"
 DESC="Extracts downloads so Radarr or Sonarr may import them."
-GOLANGCI_LINT_ARGS="--enable-all"
+GOLANGCI_LINT_ARGS="--enable-all -D dupl"
 # Example must exist at examples/$CONFIG_FILE.example
 CONFIG_FILE="unpackerr.conf"
 LICENSE="MIT"
